pyproject.toml
[project]
name = "careerai"
version = "0.1.0"
description = ""
authors = [
    {name = "Sejal Parate",email = "sejbp3003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
